requirements.txt
websocket-client
websocket-server
Flask
websockets
rich
art==5.9
uploadserver